Move things around so mu-debug.py only depends on stdlib (ie not on any PyQt imports within Mu itself)
