+++ mu/debugger/__init__.py
@@ -1,0 +1,2 @@
+# Port number for debugger.
+DEBUGGER_PORT = 31415
